Javadocs added
